// src/app/Feature/Motobike/models/add-motorbike.model.ts
export interface AddMotorbike{
	Name: string
	  Type: number 
	  Speed: number | null
	  Capacity: number | null
	  Color: string
	  YearOfManufacture: number | null
	  MadeIn: string
	  status: number
	  Description: string
	  PriceDay: number | null
	  PriceWeek: number | null
	  PriceMonth: number | null
	  LicensePlate: string
	  Image: File | null
	  CompanyName: string
  }
  